backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg
alembic==1.13.3
pydantic==2.9.2
httpx==0.27.2
boto3==1.35.36
requests==2.32.3
tiktoken
python-dotenv==1.0.1
celery==5.4.0
redis==5.1.1
torch
transformers
